Made the demo "universal"
- Modified `TodayCollectionViewController.swift` so it calculates the
sizes of the elements dynamically trying to match the ideal width 335,
on iPhones is 1 column portrait, 2 landscape on iPads is 3 and 4 or 4
and 5 depending the device. (Go ahead, try side by side apps on the
iPad, IT WORKS BABY!)
- Added an object to track the location of the cell if it changes due
to a screen rotation, otherwise during the “dismiss” it would animate
to the previous position and size and it wouldn’t match
- Changed the constraints in the card view so the image is centered and
upscaled (but never downscaled), as a result of this, the ratio
property is not longer necessary.
- Added constraints to keep the blur at the same size as the container,
otherwise presenting in one orientation and dismissing in a different
orientation or screen size (side by side apps) would result in the blur
not covering the entire screen.
- Added a new convenience initializer to
`NSEdgeLayoutConstraints.swift` but is not working as I hoped :(